Move `openSidePanel` to `sidePanelMigration`
